// src/global.enums.ts
export enum BladeburnerSkillName {
    BladesIntuition = "Blade's Intuition",
    Cloak = "Cloak",
    ShortCircuit = "Short-Circuit",
    DigitalObserver = "Digital Observer",
    Tracer = "Tracer",
    Overclock = "Overclock",
    Reaper = "Reaper",
    EvasiveSystem = "Evasive System",
    Datamancer = "Datamancer",
    CybersEdge = "Cyber's Edge",
    HandsOfMidas = "Hands of Midas",
    Hyperdrive = "Hyperdrive",
}